requirements.txt
 setuptools
requests
paramiko
python-jenkins
scp
dnspython
PyYAML
.



